Refactor functions out from routes
