+++ dataplane-webserver/src/routes/secrets.rs
@@ -1,17 +1,12 @@
-use crate::config;
-use actix_web::{get, web, Error, HttpRequest, HttpResponse};
-use k8s_openapi::ByteString;
-use kube::{Api, Client};
-use lazy_static::lazy_static;
-use log::error;
-
 use crate::secrets::types::AvailableSecret;
-use std::collections::BTreeMap;
+use crate::{config, secrets};
+use actix_web::{get, web, Error, HttpRequest, HttpResponse};
+use lazy_static::lazy_static;
+use log::warn;
 use std::ops::Deref;
-use utoipa::ToSchema;
 
 lazy_static! {
-    static ref SECRETS_ALLOW_LIST: Vec<AvailableSecret> = {
+    pub static ref SECRETS_ALLOW_LIST: Vec<AvailableSecret> = {
         let mut secrets_allow_list: Vec<AvailableSecret> = Vec::new();
         secrets_allow_list.push(
             AvailableSecret {
@@ -87,105 +82,13 @@ pub async fn get_secret(
 ) -> Result<HttpResponse, Error> {
     let (namespace, secret_name) = path.into_inner();
 
-    // Find the appropriate Secret configuration
-    let secret_config = SECRETS_ALLOW_LIST
-        .iter()
-        .find(|&secret| secret.name == secret_name);
-
-    if secret_config.is_none() {
-        return Ok(HttpResponse::NotFound().json(format!(
-            "Secret '{}' not found as allowed name",
-            secret_name
-        )));
-    }
-    let secret_config = secret_config.expect("We just checked this is not none");
-    let kubernetes_secret_name = secret_config.kube_secret_name(&namespace);
-
-    let kubernetes_client = match Client::try_default().await {
-        Ok(client) => client,
-        Err(_) => {
-            error!("Failed to create Kubernetes client");
-            return Ok(
-                HttpResponse::InternalServerError().json("Failed to create Kubernetes client")
-            );
+    let requested_secret = match secrets::validate_requested_secret(&secret_name) {
+        Ok(secret) => secret,
+        Err(e) => {
+            warn!("Invalid secret requested: {}", secret_name);
+            return Ok(HttpResponse::Forbidden().json(e));
         }
     };
 
-    let secrets_api: Api<k8s_openapi::api::core::v1::Secret> =
-        Api::namespaced(kubernetes_client, &namespace);
-    let kube_secret = secrets_api.get(&kubernetes_secret_name).await;
-
-    match kube_secret {
-        Ok(secret) => {
-            let mut filtered_data: BTreeMap<String, String> = BTreeMap::new();
-            let secret_data = match secret.data {
-                None => {
-                    error!(
-                        "Secret '{}' found in namespace '{}' does not have a 'data' block.",
-                        kubernetes_secret_name, namespace
-                    );
-                    return Ok(HttpResponse::NotFound().json(format!(
-                        "Secret '{}' not found to have data block",
-                        secret_name
-                    )));
-                }
-                Some(data) => data,
-            };
-            for key in &secret_config.possible_keys {
-                if let Some(value) = secret_data.get(key) {
-                    let value = match byte_string_to_string(value) {
-                        Ok(value) => value,
-                        Err(response) => return Ok(response),
-                    };
-                    filtered_data.insert(key.clone(), value);
-                }
-            }
-            Ok(HttpResponse::Ok().json(filtered_data))
-        }
-        Err(_) => {
-            error!(
-                "Secret '{}' not found in namespace '{}'",
-                kubernetes_secret_name, namespace
-            );
-            Ok(HttpResponse::NotFound().json(format!("Secret '{}' not found", secret_name)))
-        }
-    }
-}
-
-fn byte_string_to_string(byte_string: &ByteString) -> Result<String, HttpResponse> {
-    match String::from_utf8(byte_string.0.clone()) {
-        Ok(value) => Ok(value),
-        Err(_) => {
-            error!("Failed to convert secret value to UTF-8 string");
-            Err(HttpResponse::InternalServerError()
-                .json("Failed to convert secret value to UTF-8 string"))
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::routes::secrets::byte_string_to_string;
-    use k8s_openapi::api::core::v1::Secret;
-
-    #[test]
-    fn test_byte_string_to_string_from_json() {
-        let k8s_secret_data = r#"
-        {
-          "apiVersion": "v1",
-          "kind": "Secret",
-          "metadata": {
-            "name": "my-secret"
-          },
-          "data": {
-            "username": "dXNlcm5hbWU="
-          }
-        }
-        "#;
-        let secret: Secret = serde_json::from_str(k8s_secret_data).unwrap();
-        let secret_data = secret.data.unwrap();
-        let username_byte_string = secret_data.get("username").unwrap();
-        let result = byte_string_to_string(username_byte_string).unwrap();
-        assert_eq!(result, "username");
-    }
+    Ok(secrets::get_secret_data_from_kubernetes(namespace, requested_secret).await)
 }
